requirements.txt
torch
torchvision
matplotlib
tqdm
PyYAML
torchmetrics
torch-fidelity
numpy
